controller works
